[ros] new separate ros module
But this doesnt quite gel since there is an existing subscribers module too
